+++ email_auth/interfaces/rest/serializers.py
@@ -1,11 +1,17 @@
+import logging
 from typing import Optional
 
 import email_utils
 from django.conf import settings
+from django.contrib.auth import password_validation
+from django.core.exceptions import ValidationError
 from django.utils.translation import ugettext_lazy as _
 from rest_framework import serializers
 
 from email_auth import models
+
+
+logger = logging.getLogger(__name__)
 
 
 class EmailVerificationRequestSerializer(serializers.Serializer):
@@ -106,3 +112,93 @@ class EmailVerificationSerializer(serializers.Serializer):
             )
 
         return token
+
+
+class PasswordResetRequestSerializer(serializers.Serializer):
+    """
+    Serializer to create and send a password reset token to a verified
+    email address.
+    """
+
+    email = serializers.EmailField()
+
+    def save(self, **kwargs) -> Optional[models.PasswordReset]:
+        """
+        Send a new password reset token to the provided email address if
+        the email has already been verified. If the provided email has
+        not been verified, no action is taken.
+
+        Returns:
+            The created :py:class:`PasswordReset` instance if one was
+            created or else ``None``.
+        """
+        try:
+            email = models.EmailAddress.objects.get(
+                address__iexact=self.validated_data["email"], is_verified=True
+            )
+        except models.EmailAddress.DoesNotExist:
+            return None
+
+        reset = models.PasswordReset(email=email)
+        reset.send_email()
+
+        return reset
+
+
+class PasswordResetSerializer(serializers.Serializer):
+    """
+    Serializer to reset a user's password.
+    """
+
+    password = serializers.CharField(
+        style={"input_type": "password"}, write_only=True
+    )
+    token = serializers.CharField(
+        max_length=models.TOKEN_LENGTH, write_only=True
+    )
+
+    _reset: models.PasswordReset = None
+
+    def save(self, **kwargs):
+        """
+        Reset the password of the user associated with the provided
+        password reset token.
+        """
+        reset = self._reset
+        user = reset.email.user
+        user.set_password(self.validated_data["password"])
+        user.save()
+
+        logger.info("Reset the password for user %r", user)
+
+        reset.delete()
+
+    def validate(self, attrs: dict) -> dict:
+        """
+        Ensure that the provided token is valid and the provided
+        password passes Django's built in password validation.
+
+        Args:
+            attrs:
+                The data to validate.
+
+        Returns:
+            The validated data.
+        """
+        try:
+            self._reset = models.PasswordReset.objects.get(
+                token=attrs["token"]
+            )
+        except models.PasswordReset.DoesNotExist:
+            raise serializers.ValidationError(
+                {"token": _("The provided password reset token is invalid.")}
+            )
+
+        try:
+            password_validation.validate_password(
+                attrs["password"], user=self._reset.email.user
+            )
+        except ValidationError as e:
+            raise serializers.ValidationError({"password": e.messages})
+
+        return attrs